requirements.txt
paho-mqtt
requests
matplotlib
